# improve: GetAllCharsets and GetCharsetInfoByID