### What problem does this PR solve? 
https://github.com/pingcap/tidb/issues/9635
some of the functions may just return constants. There is no need to do compute every time if we save a map/slice and initialize them. Such as: GetAllCharsets and GetCharsetInfoByID 

### What is changed and how it works?

1. GetAllCharsets is inited at the init function, and GetAllCharsets can return the inited array directly.
2. GetCharsetInfoByID use a map to fetch Collation effectively, init the map in the init function too.


### Check List <!--REMOVE the items that are not applicable-->
 - Unit test


Code changes
 - improve GetAllCharsets and GetCharsetInfoByID
 
Related changes
 - Need to update go.mod in tidb