@shenli 2019-03-18
@hawkingrei Thanks!
Is there any benchmark result to show the improvement?