@hawkingrei 2019-03-18
BenchmarkGetAllCharsets

Before improve:

```
go test -bench Benchmark -run GetAllCharsets ./... -count 5

goos: darwin
goarch: amd64
pkg: github.com/pingcap/parser/charset
BenchmarkGetAllCharsets-4   	 3000000	       405 ns/op
BenchmarkGetAllCharsets-4   	 3000000	       405 ns/op
BenchmarkGetAllCharsets-4   	 3000000	       397 ns/op
BenchmarkGetAllCharsets-4   	 3000000	       447 ns/op
BenchmarkGetAllCharsets-4   	 3000000	       443 ns/op
PASS
ok  	github.com/pingcap/parser/charset	8.616s
PASS
```

After improve
```
go test -bench Benchmark -run GetAllCharsets ./... -count 5

goos: darwin
goarch: amd64
pkg: github.com/pingcap/parser/charset
BenchmarkGetAllCharsets-4   	2000000000	         0.38 ns/op
BenchmarkGetAllCharsets-4   	2000000000	         0.37 ns/op
BenchmarkGetAllCharsets-4   	2000000000	         0.37 ns/op
BenchmarkGetAllCharsets-4   	2000000000	         0.36 ns/op
BenchmarkGetAllCharsets-4   	2000000000	         0.36 ns/op
PASS
```

Here is benchmark code.

```
func BenchmarkGetAllCharsets(b *testing.B) {
	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		GetAllCharsets()
	}
}
```