@hawkingrei 2019-03-18
Benchmark GetAllCharsets

Before improve:

```
go test -bench Benchmark -run GetGetCharsetInfoByID ./... -count 2

goos: darwin
goarch: amd64
pkg: github.com/pingcap/parser/charset
BenchmarkGetCharsetInfoByID-4   	30000000	        64.4 ns/op
BenchmarkGetCharsetInfoByID-4   	30000000	        72.4 ns/op
PASS
ok  	github.com/pingcap/parser/charset	4.208s
```

After improve

```
go test -bench Benchmark -run GetGetCharsetInfoByID ./... -count 2

goos: darwin
goarch: amd64
pkg: github.com/pingcap/parser/charset
BenchmarkGetCharsetInfoByID-4   	100000000	        20.6 ns/op
BenchmarkGetCharsetInfoByID-4   	100000000	        18.5 ns/op
PASS
```

Here is benchmark code.

```
func BenchmarkGetCharsetInfoByID(b *testing.B) {
	index := rand.Intn(248)
	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		GetCharsetInfoByID(index)
	}
}
```