@zimulala 2019-03-19
/run-all-tests